requirements.txt
Flask
Cython==0.29.34
nemo-toolkit[all]==1.18.0
